+++ molo/surveys/wagtail_hooks.py
@@ -10,6 +10,9 @@ from django.utils.translation import ugettext as _
 
 from wagtail.wagtailadmin.menu import MenuItem
 
+from molo.surveys.models import MoloSurveyPage, SurveyTermsConditions
+from molo.core.models import ArticlePage
+
 from .views import create
 
 
@@ -50,3 +53,22 @@ def register_csv_group_creation_menu_item():
         _('CSV group creation'),
         reverse('csv-group-creation'),
         classnames='icon icon-group', order=601)
+
+
+@hooks.register('after_copy_page')
+def create_new_page_relations(request, page, new_page):
+    if page and new_page:
+        if new_page.get_descendants().count() >= \
+                page.get_descendants().count():
+            for survey in MoloSurveyPage.objects.descendant_of(
+                    new_page.get_site().root_page):
+                # replace old terms and conditions with new one, if it exists
+                relations = SurveyTermsConditions.objects.filter(page=survey)
+                for relation in relations:
+                    if relation.terms_and_conditions:
+                        new_article = ArticlePage.objects.descendant_of(
+                            new_page.get_site().root_page).filter(
+                                slug=relation.terms_and_conditions.slug)\
+                            .first()
+                        relation.terms_and_conditions = new_article
+                        relation.save()