+++ CHANGES.rst
@@ -1,7 +1,20 @@
 CHANGE LOG
 ==========
 
-5.4.0
+5.6.2
+-----
+- Bug Fix: remove PreventDeleteMixin from Ts&Cs index page
+
+5.6.1
+-----
+- Use FooterPage instead of ArticlePage for the Surveys Ts&Cs
+
+5.6.0
+-----
+- Added Terms and Conditions index page and page relation to molo survey page
+- Added image and body content to survey
+
+5.5.0
 -----
 - Add advanced surveys